+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ### Unreleased
+#### Added
+- Added automated testing for select response events ([#40](https://github.com/laurenball/Steven/pull/40))
+
 #### Fixed
 - Fixed issue with `!display` command not working for users with no configured actions ([#36](https://github.com/laurenball/Steven/pull/36))